Refactor out confirmOverwrite, PR fixes
