+++ src/utils/workspace.ts
@@ -4,8 +4,10 @@
  *--------------------------------------------------------------------------------------------*/
 
 import * as path from 'path';
+import { extensionPrefix } from 'src/constants';
+import { ext } from 'src/extensionVariables';
 import * as vscode from 'vscode';
-import { IAzureQuickPickItem, IAzureUserInput } from 'vscode-azureextensionui';
+import { IAzureQuickPickItem, IAzureUserInput, TelemetryProperties, UserCancelledError } from 'vscode-azureextensionui';
 
 export async function selectWorkspaceFile(ui: IAzureUserInput, placeHolder: string, getSubPath?: (f: vscode.WorkspaceFolder) => string | undefined): Promise<string> {
     let defaultUri: vscode.Uri | undefined;
@@ -48,4 +50,49 @@ export async function selectWorkspaceItem(ui: IAzureUserInput, placeHolder: stri
     }
 
     return folder && folder.data ? folder.data : (await ui.showOpenDialog(options))[0].fsPath;
+}
+
+export async function showWorkspaceFoldersQuickPick(placeHolderString: string, telemetryProperties: TelemetryProperties, subPathSetting: string | undefined): Promise<string> {
+    const folderQuickPickItems: IAzureQuickPickItem<string | undefined>[] = vscode.workspace.workspaceFolders ? vscode.workspace.workspaceFolders.map((value) => {
+        {
+            let fsPath: string = value.uri.fsPath;
+            if (subPathSetting) {
+                const subpath: string | undefined = vscode.workspace.getConfiguration(extensionPrefix, value.uri).get(subPathSetting);
+                if (subpath) {
+                    fsPath = path.join(fsPath, subpath);
+                }
+            }
+
+            return {
+                label: path.basename(fsPath),
+                description: fsPath,
+                data: fsPath
+            };
+        }
+    }) : [];
+
+    folderQuickPickItems.push({ label: '$(file-directory) Browse...', description: '', data: undefined });
+
+    const folderQuickPickOption = { placeHolder: placeHolderString };
+    telemetryProperties.cancelStep = 'showWorkspaceFolders';
+    const pickedItem = await ext.ui.showQuickPick(folderQuickPickItems, folderQuickPickOption);
+    telemetryProperties.cancelStep = '';
+
+    if (!pickedItem.data) {
+        const browseResult = await vscode.window.showOpenDialog({
+            canSelectFiles: false,
+            canSelectFolders: true,
+            canSelectMany: false,
+            defaultUri: vscode.workspace.workspaceFolders ? vscode.workspace.workspaceFolders[0].uri : undefined
+        });
+
+        if (!browseResult) {
+            telemetryProperties.cancelStep = 'showWorkspaceFoldersBrowse';
+            throw new UserCancelledError();
+        }
+
+        return browseResult[0].fsPath;
+    } else {
+        return pickedItem.data;
+    }
 }
